backup_20250807_004837/requirements.txt
RPi.GPIO==0.7.1
pygame==2.1.0
pydub==0.25.1
Pillow==9.5.0
luma.core==2.4.2
luma.oled==3.12.0
smbus2==0.4.2